LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/revert-layer-008-expected.txt


FAIL 'revert-layer' should revert font-size to 20px and trigger a smooth transition assert_equals: expected "15px" but got "10px"
